// frontend/src/app/api/model/taskCreate.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TaskPriority } from './taskPriority';
import { TaskStatus } from './taskStatus';


export interface TaskCreate { 
    title: string;
    description?: string | null;
    /**
     * Default TODO
     */
    status?: TaskStatus;
    /**
     * Default MEDIUM
     */
    priority?: TaskPriority;
    dueDate?: string | null;
}
export namespace TaskCreate {
}