// frontend/src/app/api/model/taskUpdate.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TaskPriority } from './taskPriority';
import { TaskStatus } from './taskStatus';


export interface TaskUpdate { 
    title: string;
    description?: string | null;
    status: TaskStatus;
    priority: TaskPriority;
    dueDate?: string | null;
}
export namespace TaskUpdate {
}